requirements.txt
IPython
matplotlib==3.10.3
gym==0.26.2
gym_sokoban==0.0.6
peft==0.16.0
accelerate==1.9.0
codetiming==1.4.0
datasets==4.0.0
dill==0.3.8
# flash-attn==2.7.4.post1
hydra-core==1.3.2
numpy==1.26.4
pandas==2.3.1
pybind11==3.0.0
ray>=2.10
tensordict>=0.8.0,<0.9.0 # verl 0.2.0.dev0 requires tensordict<=0.6.2, which is not installed.
transformers==4.53.2
vllm==0.8.2
wandb==0.21.0
gymnasium==1.2.0
gymnasium[toy-text]

pyarrow>=15.0.0
pylatexenc==2.10
torchdata==0.11.0
debugpy==1.8.15

together==1.5.21
anthropic==0.58.2

tensorboard==2.20.0
sentence-transformers==5.1.0
fast-bleu==0.0.90
nltk==3.9.1

# huggingface-hub=0.33.4

# This is optional, but needs to be installed with main requirements if you want to use webshop
# -r external/webshop-minimal/requirements.txt
